# LRUCache improvements