--- a/include/IECore/LRUCache.h
+++ b/include/IECore/LRUCache.h
@@ -35,9 +35,6 @@
 #ifndef IECORE_LRUCACHE_H
 #define IECORE_LRUCACHE_H
 
-#include "tbb/spin_mutex.h"
-#include "tbb/concurrent_unordered_map.h"
-
 #include "boost/noncopyable.hpp"
 #include "boost/function.hpp"
 #include "boost/variant.hpp"
@@ -45,36 +42,55 @@
 namespace IECore
 {
 
+namespace LRUCachePolicy
+{
+
+/// Not threadsafe. Either use from only a single thread
+/// or protect with an external mutex. Key type must have
+/// a `hash_value` implementation as described in the boost
+/// documentation.
+template<typename LRUCache>
+class Serial;
+
+/// Threadsafe, `get()` blocks if another thread is already
+/// computing the value. Key type must have a `hash_value`
+/// implementation as described in the boost documentation.
+template<typename LRUCache>
+class Parallel;
+
+} // namespace LRUCachePolicy
+
 /// A mapping from keys to values, where values are computed from keys using a user
 /// supplied function. Recently computed values are stored in the cache to accelerate
 /// subsequent lookups. Each value has a cost associated with it, and the cache has
 /// a maximum total cost above which it will remove the least recently accessed items.
 ///
-/// The Key type must have a tbb_hasher implementation.
-///
 /// The Value type must be default constructible, copy constructible and assignable.
 /// Note that Values are returned by value, and erased by assigning a default constructed
 /// value. In practice this means that a smart pointer is the best choice of Value.
+///
+/// The Policy determines the thread safety, eviction and performance characteristics
+/// of the cache. See the documentation for each individual policy in the LRUCachePolicy
+/// namespace.
 ///
 ///	The GetterKey may be used where the GetterFunction requires some auxiliary information
 /// in addition to the Key. It must be implicitly castable to Key, and all GetterKeys
 /// which yield the same Key must also yield the same results from the GetterFunction.
 ///
-/// \threading It is safe to call the methods of LRUCache from concurrent threads.
 /// \ingroup utilityGroup
-template<typename Key, typename Value, typename GetterKey=Key>
+template<typename Key, typename Value, template <typename> class Policy=LRUCachePolicy::Parallel, typename GetterKey=Key>
 class LRUCache : private boost::noncopyable
 {
 	public:
 
 		typedef size_t Cost;
+		typedef Key KeyType;
 
 		/// The GetterFunction is responsible for computing the value and cost for a cache entry
 		/// when given the key. It should throw a descriptive exception if it can't get the data for
-		/// any reason. It is unsafe to access the LRUCache itself from the GetterFunction.
+		/// any reason.
 		typedef boost::function<Value ( const GetterKey &key, Cost &cost )> GetterFunction;
 		/// The optional RemovalCallback is called whenever an item is discarded from the cache.
-		///  It is unsafe to access the LRUCache itself from the RemovalCallback.
 		typedef boost::function<void ( const Key &key, const Value &data )> RemovalCallback;
 
 		LRUCache( GetterFunction getter );
@@ -125,6 +141,9 @@ class LRUCache : private boost::noncopyable
 		// Data
 		//////////////////////////////////////////////////////////////////////////
 
+		// Give Policy access to CacheEntry definitions.
+		friend class Policy<LRUCache>;
+
 		// A function for computing values, and one for notifying of removals.
 		GetterFunction m_getter;
 		RemovalCallback m_removalCallback;
@@ -138,27 +157,9 @@ class LRUCache : private boost::noncopyable
 		};
 
 		// The type used to store a single cached item.
-		struct CacheEntry;
-
-		// Map from keys to items - this forms the basis of
-		// our cache. The concurrent_unordered_map has the
-		// following pertinent properties :
-		//
-		// - find(), insert() and traversal may be invoked concurrently
-		// - erase() may _not_ be invoked concurrently (hence we don't use it)
-		// - value access is not locked in any way (we are responsible for
-		// our own locking when accessing the CacheEntry).
-		typedef tbb::concurrent_unordered_map<Key, CacheEntry> Map;
-		typedef typename Map::iterator MapIterator;
-		typedef typename Map::const_iterator ConstMapIterator;
-		typedef typename Map::value_type MapValue;
-		Map m_map;
-
-		// CacheEntry implementation - a single item of the cache.
 		struct CacheEntry
 		{
-			CacheEntry(); // status == Uncached, previous == next == NULL
-			CacheEntry( const CacheEntry &other );
+			CacheEntry();
 
 			// We use a boost::variant to compactly store
 			// a union of the data needed for each Status.
@@ -171,85 +172,30 @@ class LRUCache : private boost::noncopyable
 			State state;
 			Cost cost; // the cost for this item
 
-			// Pointers to previous and next items
-			// in the LRU list. We use the CacheEntries
-			// themselves to store the list because it is
-			// quicker than using an external structure
-			// like a std::list. Note that although the
-			// purpose of the list is to track items where
-			// status==Cached, the two are not updated
-			// atomically, so it may _not_ be assumed
-			// that status==Cached implies previous!=NULL
-			// or status!=Cached implies previous==NULL
-			// at any given moment.
-			MapValue *previous;
-			MapValue *next;
-
-			// Mutex - must be held before accessing any
-			// fields other than the list fields (previous
-			// and next). To access the list fields, m_listMutex
-			// must be held instead.
-			tbb::spin_mutex mutex;
-
 			Status status() const;
 
 		};
 
-		// Dummy MapValues to represent the start and end of our LRU list.
-		// Items are moved to the end when they are accessed, and removed
-		// from the start when we need to reduce costs.
-		MapValue m_listStart;
-		MapValue m_listEnd;
+		// Policy. This is responsible for
+		// the internal storage for the cache.
+		Policy<LRUCache> m_policy;
 
-		// The list is inherently a serial data structure, so we must
-		// protect all accesses with this mutex. The mutex _must_ be held
-		// before the list fields of _any_ MapValue may be accessed.
-		typedef tbb::spin_mutex ListMutex;
-		ListMutex m_listMutex;
-
-		// Total cost. We store the current cost atomically so it can be updated
-		// concurrently by multiple threads.
-		typedef tbb::atomic<Cost> AtomicCost;
-		AtomicCost m_currentCost;
 		Cost m_maxCost;
 
 		// Methods
-		//
-		// Note that great care must be taken to properly handle the
-		// CacheEntry and list mutexes to avoid deadlock. If both m_listMutex
-		// and a CacheEntry::mutex must be held, the list mutex must be acquired
-		// _first_, and the CacheEntry::mutex _second_. Pay attention to the
-		// documentation for each method, to ensure that the right locks are held
-		// at the right times.
-		//////////////////////////////////////////////////////////////////////////
+		// =======
 
-		// Updates the cache entry with the new value and updates m_currentCost
-		// appropriately. Returns true if the value was stored and false if it
-		// exceeded the maximum cost. The mutex for the CacheEntry must be held by
-		// the caller.
-		bool setInternal( MapValue *mapValue, const Value &value, Cost cost );
+		// Updates the cached value and updates the current
+		// total cost.
+		bool setInternal( const Key &key, CacheEntry &cacheEntry, const Value &value, Cost cost );
 
-		// Sets the status for the cache entry to Uncached, removes any
-		// previously Cached value, updates m_currentCost and removes
-		// the entry from the LRU list. The caller must hold m_listMutex, and
-		// must _not_ hold the mutex for the cache entry.
-		bool eraseInternal( MapValue *mapValue );
+		// Removes any cached value and updates the current total
+		// cost.
+		bool eraseInternal( const Key &key, CacheEntry &cacheEntry );
 
-		// Caller must not hold any locks.
-		void limitCost();
-
-		// Either erases the item from the list, or moves it to
-		// the end, depending on whether or not it is cached.
-		// Caller must not hold any locks.
-		void updateListPosition( MapValue *mapValue );
-
-		// If the item is in the list, erases it, otherwise
-		// does nothing. Caller must hold m_listMutex.
-		void listErase( MapValue *mapValue );
-		// Inserts the item at the end of the list - the
-		// item must _not_ already be in the list.
-		// Caller must hold m_listMutex.
-		void listInsertAtEnd( MapValue *mapValue );
+		// Removes items from the cache until the current cost is
+		// at or below the specified limit.
+		void limitCost( Cost cost );
 
 		static void nullRemovalCallback( const Key &key, const Value &value );