… src/IECoreGL/CachedConverter.cpp
… test/IECore/CachedReader.py
… test/IECore/ComputationCacheTest.cpp
… test/IECore/LRUCacheTest.py
… test/IECore/ObjectPoolTest.py