This rewrites the LRUCache to provide :

- Improved performance
- Separate policies optimised for single threaded and multithreaded use cases
- Improved error handing - no more "previous attempt to get item failed"
- Optional GetterKey template argument for passing auxiliary information to the getter function
- Support for recursive calls back into `cache.get()` from the getter function itself

The performance of the previous cache was poor enough that Gaffer has been using a [modified version](https://github.com/GafferHQ/gaffer/blob/master/include/Gaffer/Private/IECorePreview/LRUCache.h) for a long time now, with the intention of moving it into Cortex. This PR is _not_ that same version though - it lacked everything here except the improved performance, and crucially lacked the support for recursive calls, which are essential for future work we want to do in Gaffer.

I've been testing this by timing `IECore.testLRUCacheThreading` in four distinct usage patterns :

- Contended on 1 : All threads hammering a single item in the cache. This is a bit of a worst case scenario for managing the locking efficiently.
- Working set 100 : A working set of 100 items, with all threads continuously requesting random items. This provides opportunities for reducing contention between threads.
- No duplicates : Pulling 10000000 unique items out of the cache, with no item requested twice. The cache is big enough to store everything without evictions, but we gain no benefit from the caching at all.
- Thrashing : Pulling random items out of a set of 1000 out of a cache with a size limit of 2.

This gives the following timings (best of 3 runs in each case, smaller numbers better) :

| | Master | This PR |  Gaffer |
| --- | --- | --- | --- |
|Contended on 1| 7.40 | 1.28 | 1.04 |
|Working set 100| 4.36 | 0.53 | 0.30 |
|No duplicates| 11.7 | 8.8 | 7.2| 
|Thrashing| 0.11 | 0.34 | 0.23 |

So for the scenarios in which we have sufficient cache hits (the first two), the new version gives us a 5-8x speedup over the original. The Gaffer version is even better (7-14x), but lacks the recursive property that we need for the future, which rules it out. Both newer versions suffer for the thrashing case, but I think it's more important that they perform well for hits.

The Serial policy provides improved performance for when the cache will be used from a single thread so no locking is required - somewhere between 2-4x the speed of the Parallel policy :

| | Parallel | Serial |
|---|---|---|
|Contended on 1|  0.87 | 0.39  |
|Working set 100| 0.87 | 0.38 |
|No duplicates| 8.76 | 2.27 | 
|Thrashing| 0.05 | 0.015 |

I've no doubt this isn't the last word in caching, but it does represent a big improvement over the status quo. I was really hoping to get back to the same performance as the Gaffer version, but despite banging my head against it for a long time I've failed. If anyone with a fresh set of eyes and some bright ideas wants to take a stab at further improvements still, that would be brilliant...