@andrewkaufman 2017-07-26
> The Gaffer version is even better (7-14x), but lacks the recursive property that we need for the future, which rules it out

Can you elaborate to refresh my memory? What does the future hold?