@johnhaddon 2017-07-26
> Can you elaborate to refresh my memory? What does the future hold?

At present in Gaffer, computation for a cache entry is non-blocking. If two threads want the same thing and it's not in the cache (as determined using `cache.get()` and a null getter function), they'll both compute it and the first to finish will update the cache (using an explicit call to `cache.set()`) That was an OK (ish!) strategy when our graph topologies were simple and most computations were per-location and didn't depend on other locations. But now we have quite a few graph patterns where many computations will depend on a single one, which in turn depends on many others :

```
o o o
\ | /
  o
/ | \
o o o
```

So we want to use the natural blocking the cache would provide if we used a non-null getter that performed the computation in the getter itself (no more calls to `set()`). This in turn means that the getter much be able to reenter the cache to pull out cached values from upstream computations, which is what the recursion tests are all about.