@johnhaddon 2017-07-28
Should have said yesterday - I pushed a fix for a compilation error and one further fix for the tests, and Travis is happy once again...